+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `cursor` when used `poster` attribut on Video.
+
 ## [2.4.0] - 2019-04-25
 
 Features: